pyproject.toml
[project]
name = "kissaten"
version = "0.1.0"
description = "Coffee bean database and search application with worldwide roaster scrapers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    "httpx>=0.24.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "duckdb>=0.9.0",
    "fastapi[all]>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "playwright>=1.37.0",
    "python-dateutil>=2.8.0",
    "pydantic-ai>=0.0.14",
    "google-genai>=0.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "coverage>=7.3.0",
    "ruff>=0.0.287",
    "pre-commit>=3.4.0",
]

[project.scripts]
kissaten = "kissaten.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
